Create clean architecture
