+++ envergo/moulinette/regulations/conditionnalitepac.py
@@ -5,7 +5,7 @@ from django import forms
 from django.conf import settings
 
 from envergo.evaluations.models import RESULTS
-from envergo.hedges.regulations import PlantationCondition, PlantationConditionMixin
+from envergo.hedges.regulations import MinLengthPacCondition, PlantationConditionMixin
 from envergo.moulinette.forms import DisplayIntegerField
 from envergo.moulinette.regulations import CriterionEvaluator
 
@@ -148,6 +148,7 @@ class Bcae8(PlantationConditionMixin, CriterionEvaluator):
     choice_label = "Conditionnalité PAC > BCAE8"
     slug = "bcae8"
     form_class = Bcae8Form
+    plantation_conditions = [MinLengthPacCondition]
 
     RESULT_MATRIX = {
         "non_soumis": RESULTS.non_soumis,
@@ -382,22 +383,3 @@ class Bcae8(PlantationConditionMixin, CriterionEvaluator):
     def get_replantation_coefficient(self):
         R = self.R_MATRIX.get(self._result_code, D("1"))
         return R
-
-    def plantation_evaluate(self, R):
-        # no R coefficient for PAC
-        hedge_data = self.catalog["haies"]
-        minimum_length_to_plant = hedge_data.lineaire_detruit_pac()
-        left_to_plant = max(
-            0,
-            minimum_length_to_plant - hedge_data.length_to_plant_pac(),
-        )
-
-        context = {
-            "minimum_length_to_plant": minimum_length_to_plant,
-            "left_to_plant": left_to_plant,
-        }
-        pac_evaluation = PlantationCondition(
-            result=(left_to_plant == 0), context=context
-        )
-
-        return [pac_evaluation]
